Split into server and client
