+++ lib/src/server/server_keepalive.dart
@@ -1,0 +1,82 @@
+import 'package:clock/clock.dart';
+
+/// Options to configure a gRPC server for receiving keepalive signals.
+class ServerKeepAliveOptions {
+  /// The maximum number of bad pings that the server will tolerate before
+  /// sending an HTTP2 GOAWAY frame and closing the transport.
+  ///
+  /// `GRPC_ARG_HTTP2_MAX_PING_STRIKES` in the docs.
+  final int? maxBadPings;
+
+  /// The minimum time that is expected between receiving successive pings.
+  ///
+  /// `GRPC_ARG_HTTP2_MIN_RECV_PING_INTERVAL_WITHOUT_DATA_MS` in the docs.
+  final Duration minIntervalBetweenPingsWithoutData;
+
+  const ServerKeepAliveOptions({
+    this.minIntervalBetweenPingsWithoutData =
+        const Duration(milliseconds: 300000),
+    this.maxBadPings = 2,
+  });
+}
+
+/// A keep alive "manager", deciding what do to when receiving pings from a
+/// client trying to keep the connection alive, based on the set
+/// [ServerKeepAliveOptions].
+class ServerKeepAlive {
+  /// What to do after receiving too many bad pings, probably shut down the
+  /// connection to not be DDoSed.
+  final Future<void> Function()? tooManyBadPings;
+
+  final ServerKeepAliveOptions options;
+
+  /// A stream of events for every time the server gets pinged.
+  final Stream<void> pingNotifier;
+
+  /// A stream of events for every time the server receives data.
+  final Stream<void> dataNotifier;
+
+  int _badPings = 0;
+  Stopwatch? _timeOfLastReceivedPing;
+
+  ServerKeepAlive({
+    this.tooManyBadPings,
+    required this.options,
+    required this.pingNotifier,
+    required this.dataNotifier,
+  });
+
+  void handle() {
+    // If we don't care about bad pings, there is not point in listening to
+    // events.
+    if (_enforcesMaxBadPings) {
+      pingNotifier.listen((_) => _onPingReceived());
+      dataNotifier.listen((_) => _onDataReceived());
+    }
+  }
+
+  bool get _enforcesMaxBadPings => (options.maxBadPings ?? 0) > 0;
+
+  Future<void> _onPingReceived() async {
+    if (_enforcesMaxBadPings) {
+      if (_timeOfLastReceivedPing == null) {
+        _timeOfLastReceivedPing = clock.stopwatch()
+          ..reset()
+          ..start();
+      } else if (_timeOfLastReceivedPing!.elapsed >
+          options.minIntervalBetweenPingsWithoutData) {
+        _badPings++;
+      }
+      if (_badPings > options.maxBadPings!) {
+        await tooManyBadPings?.call();
+      }
+    }
+  }
+
+  void _onDataReceived() {
+    if (_enforcesMaxBadPings) {
+      _badPings = 0;
+      _timeOfLastReceivedPing = null;
+    }
+  }
+}
