+++ test/client_tests/client_keepalive_manager_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.4.1 from annotations
-// in grpc/test/shared_tests/keepalive_manager_test.dart.
+// in grpc/test/client_tests/client_keepalive_manager_test.dart.
 // Do not manually edit this file.
 
 // @dart=2.19
@@ -7,7 +7,7 @@
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:mockito/mockito.dart' as _i1;
 
-import 'keepalive_manager_test.dart' as _i2;
+import 'client_keepalive_manager_test.dart' as _i2;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
