+++ test/keepalive_test.dart
@@ -1,9 +1,10 @@
 import 'dart:async';
 
 import 'package:grpc/grpc.dart';
+import 'package:grpc/src/client/client_keepalive.dart';
 import 'package:grpc/src/client/connection.dart';
 import 'package:grpc/src/client/http2_connection.dart';
-import 'package:grpc/src/shared/keepalive.dart';
+import 'package:grpc/src/server/server_keepalive.dart';
 import 'package:http2/transport.dart';
 import 'package:test/test.dart';
